+++ app/src/main/kotlin/ac/mdiq/podcini/playback/service/PlaybackService.kt
@@ -49,6 +49,9 @@ import android.view.KeyEvent
 import android.view.KeyEvent.KEYCODE_MEDIA_STOP
 import android.view.ViewConfiguration
 import androidx.annotation.RequiresPermission
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.core.app.NotificationCompat
 import androidx.media3.common.MediaItem
 import androidx.media3.common.MediaMetadata
@@ -639,11 +642,12 @@ class PlaybackService : MediaLibraryService() {
         var playbackService: PlaybackService? = null
         var mediaBrowser: MediaBrowser? = null
 
-        var isRunning: Boolean = false
+        var isRunning by mutableStateOf(false)
 
         var isCasting: Boolean = false
             internal set
 
+        // TODO: this appears not used
         var episodeChangedWhenScreenOff: Boolean = false
 
         /**